Indexing only on ValueDocument;
Added findbugs configuration for maven and findbugs.

+++ src/main/java/nl/thanod/Scan.java
@@ -36,7 +36,7 @@ public class Scan
 					co++;
 				}
 			}
-			float took = (System.nanoTime() - start) / 1000000;
+			double took = (System.nanoTime() - start) / 1000000.0;
 			System.out.println("took: " + took + "ms");
 			System.out.println("found: " + co);
 		}
